fix(inventory): allow fractional deltas when adjusting stock
AdjustStockResource.Delta was typed `int` while AdjustStockCommand's own
Delta is `decimal` — ASP.NET's JSON deserializer rejected any non-integral
value with a 400 before the command ever ran, so adjusting a product sold
by weight (X5 Bloque D) always failed. X6 Kardex's "Pérdida rápida" entry
point was the first thing to actually exercise this path with a fractional
value, surfacing it; the original per-row "Ajustar stock" button in
Inventario carried the same latent bug.

+++ Kipu.Platform/Products/Interfaces/Rest/Resources/AdjustStockResource.cs
@@ -1,3 +1,12 @@
 namespace Kipu.Platform.Products.Interfaces.Rest.Resources;
 
-public record AdjustStockResource(int WarehouseId, int Delta, string Reason);
+/// <summary>
+///     Delta is decimal, matching AdjustStockCommand — a product sold by
+///     weight (X5 Bloque D) needs a fractional adjustment (e.g. -0.5kg
+///     spillage). It was typed `int` here until X6 Kardex's "Pérdida rápida"
+///     entry point was the first thing to actually exercise this path with a
+///     fractional value, surfacing a 400 from ASP.NET's JSON deserializer
+///     rejecting a non-integral number into an int — the original per-row
+///     "Ajustar stock" button in Inventario carried the same latent bug.
+/// </summary>
+public record AdjustStockResource(int WarehouseId, decimal Delta, string Reason);
